Move derive_nullifier function into derive_nullifier.rs file
