+++ src/circuit/circuit_zsa.rs
@@ -37,15 +37,13 @@ use crate::{
 
 use super::{
     commit_ivk::{self, CommitIvkChip},
-    gadget::{
-        add_chip::AddChip, assign_free_advice, assign_is_native_asset, assign_split_flag,
-        AddInstruction,
-    },
+    gadget::{add_chip::AddChip, assign_free_advice, assign_is_native_asset, assign_split_flag},
     note_commit::NoteCommitChip,
     Circuit, OrchardCircuit, ANCHOR, CMX, CV_NET_X, CV_NET_Y, ENABLE_OUTPUT, ENABLE_SPEND,
     ENABLE_ZSA, NF_OLD, RK_X, RK_Y,
 };
 
+mod derive_nullifier;
 pub mod gadget;
 mod note_commit;
 mod value_commit_orchard;
